+++ test/eca/features/chat_test.clj
@@ -212,7 +212,7 @@ for allowed directories and then list files"
               (on-first-response-received {:type :text :text "Ok,"})
               (on-message-received {:type :text :text "Ok,"})
               (on-message-received {:type :text :text " working on it"})
-              (on-prepare-tool-call {:id "call-1" :name "list_allowed_directories" :argumentsText ""})
+              (on-prepare-tool-call {:id "call-1" :name "list_allowed_directories" :arguments-text ""})
               (on-tool-called {:id "call-1" :name "list_allowed_directories" :arguments {}})
               (on-message-received {:type :text :text "I can see: \n"})
               (on-message-received {:type :text :text "/foo/bar"})
@@ -237,7 +237,7 @@ for allowed directories and then list files"
              {:role :system :content {:type :progress :state :running :text "Generating"}}
              {:role :assistant :content {:type :text :text "Ok,"}}
              {:role :assistant :content {:type :text :text " working on it"}}
-             {:role :assistant :content {:type :toolCallPrepare :id "call-1" :name "list_allowed_directories" :argumentText nil :manual-approval false}}
+             {:role :assistant :content {:type :toolCallPrepare :id "call-1" :name "list_allowed_directories" :arguments-text "" :manual-approval false}}
              {:role :assistant :content {:type :toolCallRun :id "call-1" :name "list_allowed_directories" :arguments {} :manual-approval false}}
              {:role :assistant :content {:type :toolCalled :id "call-1" :name "list_allowed_directories" :arguments {} :outputs [{:content "Allowed directories: /foo/bar" :error false :type :text}]}}
              {:role :assistant :content {:type :text :text "I can see: \n"}}